server/requirements.txt
python-socketio
uvicorn[standard]
fastapi